% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TakifuguRubripesChromosome12}
\alias{TakifuguRubripesChromosome12}
\title{FASFA file Takifugu rubripes chromosome 12, fTakRub1.2, whole genome shotgun sequence}
\format{
Nucleotide Sequence
}
\source{
NCBI
}
\usage{
TakifuguRubripesChromosome12
}
\description{
FASFA file Takifugu rubripes chromosome 12, fTakRub1.2, whole genome shotgun sequence
}
\examples{
\dontrun{
TakifuguRubripesChromosome12
}
}
\keyword{datasets}
